Fixed the "ampersand bug" which was preventing afs filenames which contained the characters & > < ' " from downloading properly.
Changed the output of afs.php to output its javascript "Files" array in
an unencoded form... no longer escaped for URL encoding. The only
character that required escape was the ' character.

Changed javascript which used the "Files" array to do the proper
thing when displaying the file path and title, by either
encoding it for URL or leaving it alone for plain text display.